MAIL generational GC
This is the first checkin of a (mostly) working generational GC.  It still
needs work, but passes all the regressions and self compiles to fixpoint,
using the generational GC.  There is still at least one bug when compiling
-generational false.

Added two new compiler switches to control the generational GC:
	-card-size-log2 n
	-generational {true|false}

Disabled -inline-array, because GC_allocateArray is at some point going to
know about generations, and so array allocation can no longer be inlined.  At
some point, we should probably delete the code in the compiler that does
inline array allocation.

Fixed a performance bug in heap resizing that really hurts when the amount of
live data is more than half the RAM size.  The problem was that the entire
block of live data was copied from one place to another, and then the old
space was munmapped.  That meant that both copies of the live data had to
exist in virtual memory simultaneously.  Now, the copy is done from high
address to low address, 10M at a time, munmapping parts of the old space as we
go.  That way, as long as the live data fits in RAM (give or take), no
paging need occur.

Turned Thread_current from a primitive into an FFI call.