MAIL unreachable blocks bugfix
Fixed a bug in constant-propagation and useless that was triggered when they
received an input with unreachable blocks.  Both passes would fail in such a
situation because they do analysis only on reachable blocks, but rewrite based
on unreachable ones.  The fix was to run a prepass to eliminate unreachable
blocks.

There is still possibly a similar bug in local-ref.

Two other ways to fix the problem would be to

1. disallow unreachable blocks entirely.
2. change the buggy passes to do rewriting only on reachable blocks.

(1) seems pretty expensive to me in terms of compiler run time, because every
shrinker pass would have to be followed by a pass that eliminates unreachable
blocks.

(2) might be OK, but requires some more complicated rewriting of the offending
passes than what I did here.