Change the card map used by the mutator in GC_state to cardMapForMutator.
Separately allocate the card map and cross map from the rest of the heap.  This
was done in anticipation of using mremap to grow the heap, which also needs to
grow the card map and the cross map.  That would have been impossible with the
old approach where the card map and cross map were of fixed size at the front of
the heap.

Separating the card map and cross map out also meant that they aren't needed in
both heaps, and can be kept as part of GC_state rather than GC_heap.

Along the way, I also moved the nursery from GC_heap to GC_state.