MAIL -exn-history reimplemted
Changed the implementation of exception history to be completely functional.
Now, the extra field in exceptions (when compiling -exn-history true) is a
string list instead of a string list ref, and raise conses a new exception
with a new element in the list instead of assigning to the list.  This
changes the semantics of exception history (for the better) on some
programs. See regression/exnHistory3.sml for an example.  It also
significantly improves performance when compiling -exn-history true.