MAIL fixed segfault when compiling -profile time
Fixed bug in ProfileTime that caused a segfault sometimes when running
executables compiled -profile time.  I had introduced the problem with the merge
of the allocation profiling branch, where I had deleted the line that turned off
the resetting of the SIGPROF itimer, which meant that some signals were received
after the data buffer had been freed.

The fix was to add the line back to turn off the itimer.