MAIL Added -cc and -ccopt
Added the -cc and -ccopt switches, with the semantics pretty much as
was requested.  -cc sets the C compiler and -ccopt adds a switch to
the list of switches passed to gcc (-ccopt is only relevant when
compiling -native false).  The only minor weirdness that I put in is
that -cc doesn't reset the switch list if you just pass in the C
compiler without any switches.  So, "-cc gcc" will cause gcc to be
invoked with all the usual switches from bin/mlton.  But "-cc 'gcc
-DFOO'" will cause only the switch -DFOO to be passed, unless you also
use -ccopt.