MAIL another aproach to benchmark command line parsing
This approach differs from Matthew's in that there is a generic
ChoicePattern.expand function that only knows about brace-delimited
comma separated lists (e.g. {foo,bar,baz}) and it expands on a purely
textual level with no notion of tokenization.  Hopefully that makes it
more generally useful.  Now, the benchmark script doesn't attempt to
do any tokenization/parsing of the -mlton arg.  It simply does the
choice expansion and calls sh -c.

The approach supports nested {}, but not # since it has no notion of
tokenization.  But # isn't needed.  Matthew's earlier example can be
handled with

-mlton "mlton -v3 -native {true,false -cc gcc{, -ccopt -fno-strict-aliasing}{ -DDEBUG,}}"

I also changed -cc so that you can only specify the gcc executable and
no command line switches and so that -cc always resets the switch
list.  There was really no need to allow switches with -cc since we
can do them with -ccopt.