MAIL simple overflow detection elimination
Added a test to the redundant-tests pass that will sometimes eliminate
the overflow test when adding or subtracting 1.  In particular, it
will eliminate it in the following cases:

if x < y
  then ... x + 1 ...
else ... y - 1 ...

Maybe more importantly, in adding this, I noticed that there was a bug
introduced about a year ago in redundant-tests that caused it not to
run at all.  I also fixed that bug.