MAIL -basis command line option
Added -basis command line option (which subsumes the old
-use-basis-library option).  Currently, only basis-2002 and none are
supported.  To add a basis, add it to the Control.basisLibs list and
create a directory in /basis-library/libs with the same name, with
the files bind, prefix, and suffix.
World sizes for the partially elaborated basis are on par with the old system.