Yet more organizing.
Only basis-1997 regressions that fail with -basis basis-1997 are
real6, which uses the IEEEReal.float_class datatype.  The problem is
that the REAL signature refers to IEEEReal by name, so the conversion
of the IEEEReal structure would need to occur outside the Basis1997
stucture to correctly bind the float_class type within the REAL
signatures.