MAIL types for Rssa
Added types to Rssa and Machine, as discussed in earlier email.  The
basic idea is to have object types that correspond to the header
information that is available to the runtime, and sum types that are
used to represent datatypes.  See backend/machine-atoms.sig for the
new types.

There is now quite a bit of type checking that is done on Rssa, but it
is still far from type safe, partially due to lack of effort and
partially due to some unsolved problems.  A careful pass through the
type checker in rssa.fun should make stuff better.  But there are
still some casts used.  See castIsOk in machine-atoms.fun to see the
casts that are currently allowed.  Other things that remain to be
(type-)checked: liveness info, limit checks, case statements that
narrow types, globals defined before use, primapps, ...

Various primitives are now implemented as casts by SsaToRssa instead
of being implemented in the codegens.

	Byte_byteToChar
	Byte_charToByte
	C_CS_charArrayToWord8Array
	IntInf_fromVector
	IntInf_fromWord
	IntInf_toVector
	IntInf_toWord
	String_fromWord8Vector
	String_toWord8Vector
	Vector_fromArray
	Word32_fromInt
	Word32_toIntX

Of course, some of these are unsafe (e.g. IntInf_fromWord).
Vector_fromArray now changes the object header, in anticipation of
object headers including mutability information some day.

It should now be a simple matter to use headers as variant tags, but
I haven't done it yet.

Combined all the switch statements used by Rssa and Machine into a
single datatype -- see backend/switch.sig.  With that and the changes
to operands, Rssa and Machine are starting to look suspiciously
similar.  Hopefully one day we will be able to unify them.

Eliminated Array_array0, and Array_array is now used to allocate
zero-length arrays.  This required a minor change to the runtime,
since zero-length arrays now have the proper type tag instead of all
having the same one.

The backend register allocation is no longer attempts to share a
register for multiple variables.  This may cause performance problems
since the local{char,int,...}  arrays used by the native codegen to
cache real registers will no longer be as small or as densely used.

There were some pretty straightforward changes to the codegens to
keep up with the changes to MACHINE.

Added a new flag "-keep rssa" and added lots of improvements to Rssa
pretty printing.

Changes for rearrangement of sorting functions in library.