Propagate source level information about functions through to the executable.  This is in anticipation of modifying mlprof to report source information.
There are two kinds of information.

1. For each SSA function we record the source lambda that was
translated to it.  There are a couple of special cases where there is
no source lambda (main, functions that implement polymorphic
equality).

2. For each return address/stack frame, we record the SSA function
that makes the nontail call that creates the frame.

This information is all recorded in a single string in the executable,
which is stored in gcState.profileInfo at program startup.  There is a
new runtime argument, @MLton show-prof, that will print this string.
This is intended to be used by mlprof to get the above information so
that it can correlate the ticks to source positions.

I stored the info in the executable instead of using an extra file so
that we can continue to call mlprof in the same way and not have to
worry about an extra file containing the info.