MAIL source-level profiling
Second whack at source-level profiling.  Here's how it works.

Associate source positions (file and line) with functions in the
source and propagate through to the SSA.  In the closure converter,
when first creating each SSA function with associated source info si,
insert an "Enter si" profile statement at the beginning of the
function and a "Leave si" profile statement before each return, raise,
or tail call (see Ssa.Function.profile).  Then, all of the SSA
simplifier passes preserve the Enter/Leave statements.  The SSA type
checker checks that the Enter/Leave statements properly nested (in the
sense of balanced parentheses).  This required a few changes to the
SSA simplifier (see below).

The Enter/Leaves are directly translated into RSSA and preserved by
all of the RSSA passes.  At the end of the RSSA passes, the profile
pass (see backend/profile.fun) uses the Enter/Leave statements to
determine the stack of source functions at each program point.  It
does this by a simple depth-first search of the basic blocks in the
function, keeping track of the stack of source functions, pushing for
Enter and popping for Leave.  For time profiling, it inserts a
ProfileLabel statement at the beginning of each basic block and at
each point within the basic block that the source info stack changes.
For allocation profiling, it inserts assignments to
gcState.profileAllocIndex and calls to GC_ProfileAlloc_inc whenever
the source stack changes.

In addition to instrumenting the program, the profiling pass produces
four pieces of information (see Machine.ProfileInfo.t)

1. sources: SourceInfo.t vector
	This is used to share source info.  Each source info appears
	in sources once.  Other information refers to source info by
	indexing into this vector.
2. sourceSeqs: int vector vector
	This contains all of the local source stacks seen by the
	profile pass, given as vectors of indices into sources.
	Other information refers to source stacks by indexing into
	this vector
3. frameSources: int vector
	This describes the source stack that the profile pass
	inferred at each continuation.
4. labels: {label: ProfileLabel.t,
            sourceSeqsIndex: int} vector
	This is only used for time profiling.  It records the source
	stack that the the profile pass associates with each profile
	label.

All of this information is output into the C file as structs/arrays
and is stored in gcState at program initialization.  All the codegen
has to do is spit out the profileLabels.  This was another nice thing
about the new approach that should make profiling easier with new
codegens (or even with the C codegen, where it has been disabled for
ages).

At run time, time profiling first sorts the labels array in increasing
order of label, and uses that to build an array (gcState.textSources)
that maps program address to sourceSeqs index.  A unit of time
profiling data then just needs to keep an array of counts with one
entry for each source info.  When the profiling signal handler gets an
interrupt, it looks up the sourceSeqs index in textSources, then looks
up the sources index in sourceSeqs, and bumps the appropriate counter
in the profiling data.

Space profiling is similar.  The RSSA profiling pass has already
inserted the assignments to gcState.profileAllocIndex and calls to
MLton_ProfileAlloc_inc, which then has to look up the sourceSeq and
source index and bump the counter.

The data stored in an mlmon.out file is now very simple.  First off,
the file is now a text file, not a binary file.  The file contains a
flag indicating whether it is allocation or time data and then the
array of counts corresponding to the sources.  There is a new runtime
argument, show-prof, which prints out the strings in the sources
array.  mlprof uses show-prof, zips up the sources with the counts,
and has the profiling data.  There is no notion of depth for now.

All of this is how profiling of "leaf" source-level profiling works.
That is, the counts (be they clock ticks or bytes allocated) are
associated with leaves in the call graph.  Coming soon, I will add the
ability to have the (alloc or time) profiler walk the stack and bump a
counter for all the functions on the call stack.  Then, the deeper
sources in each sourceSeq and the frameSources will be used to map the
SSA call stack into the source call stack.  The mlmon file will still
be the same, the counts will just denote how often the function was on
the call stack.  This should give a clearer picture of where time is
being spent by the major functions program (kind of like our pass
timings with MLton -v2).

Now, for a few notes.

I added a new flag, -keep machine, since the Machine code is now much
more readable and has much more sensible type info.

I got tired of making repeated changes to x86codegen.h and ccodegen.h,
so I created a new file, codegen.h, that captures all the similar code
from them.

I had to make a few changes to the SSA simplifier and the Enter/Leave
checker in order match Enter/Leave statements.
1. Because removeUnused can turn a nontail call into a tail call with
	a Dead continuation but does not know to "pop" the Enter/Leave
	stack before the call, the Enter/Leave checker allows any
	stack before a Dead.  This is unlike Tail calls, where it
	requires the stack to be empty.
2. Because of 1, the contifier is not allowed to contify a function
	that is called with a Dead continuation -- the Enter/Leave
	stack might not match if it does.  Similarly for
	introduce-loops.
3. knownCase and removeUnused must create new blocks instead of
	sharing them since it can't guarantee that the profile stacks
	are the same at the shared blocks.
I could have put the changes under a test for
	!Control.profile <> Control.ProfileNone
but my goal is to have profiling have as small of an impact on the
program as possible, and so I decided to see if we can live with the
weaker optimizations in general.  The jury is still out.  None of the
benchmarks seem to have been hurt by the optimizer changes.  However,
there has been a significant self-compile slowdown.  I am
investigating.

The SSA simplifier changes tickled a bug in signal check insertion.
It had missed the possibility of looping forever via recursive
function calls instead of SSA loops.  To fix the problem, I added a
signal check on entry to each function.

The machine type checker is too slow because it does a linear lookup
of profile labels in an array.  I will fix it to use plists.

Time profiling does not yet correctly handle other labels (e.g. C
functions, GC functions).  In order to do that, I plan to have the
profiling initialization code do an nm on the program itself to find
the labels and then to add special "source" infos into the sources
array and the corresponding indices into the textSources array.