MAIL source-level profiling
Third whack at source-level profiling, including allocation and time
profiling of the stack.

Summary: it works, but it's too slow in some cases.

Details:

The goal of stack profiling (aka cumulative profiling) is to at each
point of interest (clock tick or allocation) bump a counter for each
source function that is on the stack (once per functioon).

For allocation profiling, I added C calls upon each enter and leave of
a source function.  These calls keep track of the number of
occurrences of each function on the stack, and store the current value
of totalBytesAllocated when a function is first entered, bumping the
counter by the difference in totalBytesAllocated when the function is
last left.  This works, but the overhead of the C calls can really
hurt, often by a factor of 5X-10X and sometimes by more.

For time profiling, I had the SIGPROF handler walk the stack, keeping
track of whether times each function is on the stack and bumping a
counter once per function.  This worked very well for most benchmarks,
but slowed merge down because it has very deep stacks.  It was so slow
I killed it after hours.  Here are the numbers.

run time ratio
benchmark         MLton1
barnes-hut          1.01
boyer               0.87
checksum            1.03
count-graphs        1.07
DLXSimulator        1.02
fft                 1.05
fib                 1.38
hamlet              1.04
imp-for             1.54
knuth-bendix        1.13
lexgen              1.01
life                1.18
logic               1.03
mandelbrot          1.09
matrix-multiply     1.06
md5                 1.25
merge 		       *
mlyacc              1.03
model-elimination   1.06
mpuz                1.09
nucleic             1.09
peek                4.67
psdes-random        1.14
ratio-regions       1.04
ray                 1.01
raytrace            1.00
simple              1.05
smith-normal-form   1.00
tailfib             0.81
tak                 1.42
tensor              0.96
tsp                 0.99
tyan                1.06
vector-concat       0.98
vector-rev          1.04
vliw                1.07
wc-input1           1.06
wc-scanStream       1.42
zebra               0.95
zern                0.98

The large ratio in peek is not due to the stack profiling -- it is
just due to profiling, since I saw it with stack profiling turned
off.

Unfortunately, the stack walking also makes a self-compile unbearably
slow (it did not complete in 10 hours).

After this checkin I will begin work on a new approach that I hope
will make the overhead acceptable for both stack and time profiling.


Here are some other changes:

Made implementHandlers work on Rssa instead of Ssa.  This let me
eliminate SetExnStackSlot, SetExnStackLocal, and SetSlotExnStack from
Ssa.  CheckHandlers now also works on Rssa instead of Ssa.

Eliminated SetExnStackSlot SetExnStackLocal and SetSlotExnStack from
MACHINE.  These are now implemented with arithmetic and moves.
Changed Machine.Type.Label to Machine.Type.Label of Label.t so that
the type checker can track handlers and make sure that the handler
stack is set when doing a nontail call.  This required changing
liveness to track the handler label.

Made handlers look more like continuations, with a frame layout and
the stack pointer pointing one word past them instead of at them when
raising.  This is so stack walking can work in the middle of a raise,
when the stack is set to a handler frame, not a continuation frame.