MAIL mlprof now shows call-stack graph
Added the call-stack graph to profile info and to what is displayed by
@MLton show-prof.  mlprof uses this to create the call graph (in .dot
format) with the the profiling data.  Right now, mlprof only displays
nodes above the threshold, but I'll probably add an option to display
the whole graph.

I refer to it as the call-stack graph rather than the call graph
because it describes the set of possible call stacks, and differs from
the call graph in how tail calls are handled.  For example if A
nontail calls B and B tail calls C, then the call-stack graph has
edges A->B A->C, while the call-graph would have A->B->C.

I removed a lot of old cruft from mlprof.