MAIL Profiling labels in x86-codegen
Ensure that all blocks introduced by the x86 codegen have an
associated profile label.  This eliminates an imprecision in time
profiling with the native codegen.  The one outstanding issue is the
fact that empty compensation blocks are not removed when they have
profile labels embedded.  This shouldn't be too hard to fix, and I'll
try to do so by the end of the week.