MAIL profiling: elimination of currentSource
Profiling no longer uses gcState.currentSource (which I have
eliminated) to get the current source.  Instead, it uses the top frame
on the ML stack.  This necessitates creating a stack frame for all C
calls when compiling with profiling.  GC_profile{Enter,Inc,Leave} also
use the top frame on the ML stack to know where they are.

In making this change, I had to tweak similar parts of
x86-generate-transfers.fun and c-codegen.fun.  Just a thought: maybe
it would be better to move synchronization of gcState to the backend
and make it more explicit in the Machine IL (i.e. make FlushStackTop,
FlushFrontier and friends Machine IL statements)?  This might
eliminate some duplication of knowledge of when to flush from all the
codegens.

Added a new field to frameLayouts to indicate whether the frame is
for calling a C function or an ML function.  The time profiling signal
handler uses this to decide whether or not to look at the current PC.

Made GC_state more consistent in keeping sizes of all arrays.