MAIL -profile time with the C codegen
Added support for -profile time to the C codegen.  This is not
intended to be perfect, since we can't control what gcc does with the
code.  But it can still be useful.

The idea is to emit a profile label via __asm__ whenever the Machine
code has a ProfileLabel statement.  The only (minor) complexity is
that gcc may prove that code is dead even if MLton didn't -- hence the
labels must be weak.  This also meant changing the runtime to allow
labels to be zero valued.

Nicely enough, this helped me find a couple of missing simplifications
in MLton's Prim.apply.