MAIL mlprof -graph set expressions
Added a little sublanguage for specifying sets of nodes to be
displayed in the dot call graph.  mlprof's -graph switch is used to
specify the set.  The language has regular expressions for specifying
node names, and the usual set operations (union, intersection,
complement), as well as two operations "from" and "to" for specify the
set of nodes on paths from and to a node, respectively.

See the user guide for details.

Made -thresh take a real argument instead of an integer.