MAIL using header words as variant tags
Added the ability to use header words as variant tags instead of
reserving the first word of the object as we used to do.  You can
control which technique is used with -variant {header|first-word}.
I've gone ahead and made -variant header the default since everything
seems to work.  All that was required was a couple of small changes to
the backend.

I ran all the benchmarks to compare the two approaches.  Here are the
only ones where the ratio was more than 0.05 away from 1.

MLton0 -- mlton -variant first-word
MLton1 -- mlton -variant header

run time ratio
benchmark         MLton1
boyer               0.87
hamlet              0.86
knuth-bendix        0.93
lexgen              1.06
logic               0.93
nucleic             1.07
peek                1.11

So, not too much of an improvement.

For self compiles, there was a minor improvement, cutting a little
time and about 1.5G of allocation.

-variant first-word
	MLton finished in 274.16 + 132.44 (33% GC)
	total allocated: 25,342,182,240 bytes

-variant header
	MLton finished in 261.86 + 120.96 (32% GC)
	total allocated: 23,611,188,200 bytes

One question about the native codegen: the variant tags for a
particular datatype will be consecutive integers, but not starting at
zero.  Is the codegen smart enough to do a subtract and make a jump
table?