MAIL TextIO.* now raises Io with name field
Associate a name with instreams and outstreams so that the Io
exception has a name.

Update the exception layout function used in the error mssage
pretty-printer for command lines so that the function and name are
displayed.

opefully, this means that when there are problems like Tom Murphy
encountered with /tmp being full, the error message will look
something like

openOut "/tmp/whatever.s": No space left on device