MAIL -profile-combine {false|true}
Added -profile-combine {false|true} option to mlton.  This lets one
achieve the old behavior of combining all duplicates of a function
into a single node.  I first tried to do this as a switch to mlprof,
but then learned the hard way that that can't work, at least with
-profile-stack true, since there is no reasonable way to combine stack
counts (two different copies of a function may be on the stack
simultaneously, or may not).