MAIL phantom types and directed graphs
Added phantom types to directed graphs to help in catching errors that
confuse different graphs.

As hoped, the implementation of graphs didn't change much.  The trick
of adding the wrapper declarations, like "type 'a t = t", to the end
of the module worked well.  The only difficulty was with datatypes,
which cannot be reparameterized in a similar manner.  In one case
(idomRes), I went ahead and added the phantom type variable to the
original declaration.  In the other case (LoopForest.t), I hid the
fact that it was a datatype.

Adding the phantom type variable might cause the monomorphiser to
create unnecessary duplicates.  But with a little bit of smarts it
should be able to notice that the type variable is unused and hence
ignore it.