MAIL elimination of unused type arguments
Added a pass that runs on XML before monomorphisation and eliminates
unused type arguments.  It first computes a simple fixpoint on all the
datatype declarations to determine which datatype tycon args are
actually used.  Then it does a single pass over the program to
determine which polymorphic declaration type variables are used, and
rewrites types to eliminate unused type arguments.

This pass should eliminate any spurious duplication that
monomorphisation might perform due to phantom types.  If you're able
to find a program that it misses, let me know.