MAIL heirarchichal naming of profiled functions
Added support for heirarchichal naming of profiled functions.  For
example, in the following code, g would appear as S.f.g.

structure S =
   struct
       fun f = ... fun g ...
   end

I used "." as the separator for both nesting of structures as well as
functions.  It looks OK to me.

There is also an expert mlprof option, -long-name {true|false}, to get
the old short names if you want.