MAIL calling SML from C
This checkin includes a simple mechanism for calling SML from C.
Warning: it only works with -native false for now.  The interface is
as follows.  On the SML side, we export a single function that allows
one to set the "handler" that runs when a call to SML is made from C:

	val MLton.FFI.handleCallFromC: (unit -> unit) -> unit

On the C side, we export a single function that allows us to call back
to SML:

	void MLton_callFromC ();

This interface and implementation is only intended to handle the
situation where SML is the main program and calls C functions that may
need to call back to SML, which is the case for mGTK.  It is easy
enough to use this primitive interface to implement the ability to
call multiple SML functions with wrappers around MLton_callFromC that
set a global integer and dispatch inside handleCallFromC that tests
the integer.  It is also easy enough to pass arguments to the SML
functions, again by setting globals in the C wrapper and fetching them
(with _ffi) on the SML side.  So, it is my hope that this interface
will be sufficient for the mGTK guys.

This interface is not intended to allow creation of SML libraries as
.o files.  But it's hopefully a step in that direction.

The implementation is based on Matthew's suggestion of using threads
to handle callbacks and is very similar to how MLton handles signals.
There is a new field, gcState.callFromCHandler, that points to the
thread that is to be run when C calls MLton_callFromC ().  There is
also one new primitive, Thread_returnToC, that the callFromCHandler
calls when it is done.  Calling MLton_callFromC () switches to the
callFromCHandlerThread, and then runs the SML until it calls
Thread_returnToC.  It then switches back to the thread that was
"interrupted" by the C call, and returns to C.

One major impact of allowing C to call back to SML is that it changes
the assumptions that the optimizer can make about C calls.  Namely, it
used to be a safe assumption that anything declared by _ffi didn't
modify the frontier or the stackTop and didn't GC.  Now, that is no
longer true, since any _ffi might call back to C, which could do any
of those things.  There are two ways to solve this problem.

1. If the program handles C calls, then assume that all _ffi's may
   modify the frontier or the stackTop and may GC.

2. Introduce an annotation on _ffi declarations that allows the
   programmer to indicate the functions that may callback.

3. Introduce an annotation on _ffi declarations that allows the
   programmer to indicate the functions that don't callback.

For now I've chosen (1), but only because it was easier to implement.
In looking at this, I see that we go to a lot of trouble in the
backend and codegen to keep track of all of these properties of C
functions (modifiesFrontier, mayGC, ...) in the backend so that the
codegens can optimize C calls.  It might make our lives a lot easier,
and the compiler more obviously correct, to simply make worst case
assumptions for some of these.  I don't know much that will hurt
performance.  It's probably worth running some experiments to check.

Warning: this checkin only supports calling SML from C with the C
codegen.  To get it to work with the native codegen, two things must
be done:

1. Add an implementation of MLton_callFromC to x86codegen.h.  It
   shouldn't be too hard to pattern match off of MLton_callFromC in
   ccodegen.h and main in x86codegen.h.

2. Implement C calls to Thread_returnToC in the x86 codegen.  Again,
   the semantics are in the C codegen.

Matthew, can you look into those two things?