Common argument optimization pass.
Optimizes instances of Goto transfers that pass
the same arguments to the same label; e.g.
 l1 () = [... z1 = ? ...] Goto l3(x, y, z1)
 l2 () = [... z2 = ? ...] Goto l3(x, y, z2)
 l3 (a, b, c) = ...
This code can be simplified to:
 l1 () = [... z1 = ? ...] Goto l3(z1)
 l2 () = [... z2 = ? ...] Goto l3(z2)
 l3 (c) = [a = x; b = y; ...]
which saves a number of resources: time of setting up the arguments
for the jump to l3, space (either stack or pseudo-registers) for the
arguments of l3, etc.  It may also expose some other optimizations, if
more information is known about x or y.