MAIL C codegen now puts chunks in separate files
Changed the C codegen so that each chunk appears in a separate C file
instead of putting them all in one file.  This should make compiling
large programs with the C codegen a lot faster.  This is done with the
intent of cross-compiling MLton to the Sparc using the C codegen.

This change required some modifications to the command-line processor
in main, so that it can handle the multiple C files that are generated
by the codegen.

I also improved the chunkifier so that it merges unrelated chunks
while obeying the size constraint, so that the number of chunks is
kept small.  For example, a self compile requires 151 chunks.

Here are the timings of four self compiles, for each of the
possibilities of a native-compiled (N) or C-compiled (C) mlton
compilng itself natively (N) or using the C codegen (C).


	total		 pre-codegen	Compile C and Assemble
	---------------- --------------	----------------------
CC	1275.91 +  61.06 198.81 + 48.84 1032.12 + 0.24
CN	 392.65 + 126.18 193.92 + 48.83   19.14 + 0.53
NN	 265.53 +  54.68 138.94 + 34.74	  19.24 + 0.50
NC	1224.23 +  40.60 143.38 + 36.32	1033.07 + 0.14

Here are the (text + data) sizes for the two compilers

native	7,637,478 +   843,840
C	8,472,030 + 1,150,280

So, while using the C codegen is a lot slower than using the native
codegen, it is still a viable option.