MAIL first checkin of SunOS/SPARC port
Added support for SunOS/SPARC platform using the C codegenerator.  It
passes all the regressions and a cross-self compile as well as normal
self compile on a SPARC.  I have not gone through a full bootstrap on
a SPARC because it is ridiculously slow (5 hours and counting) for the
version built with stubs to self compile.  There's still a lot of
cleanup and performance tuning left.

Here's a more detailed list of changes.

Added MLton.hostType Sun.

Changed the type of Random.{seed,useed} so that they return a word
option instead of a word.  They now return NONE if
/dev/{random,urandom} can't be read from (which may be the case on
SunOS).

Changed all shell scripts to /usr/bin/env bash.  I needed to do this
because sh on SunOS doesn't have !.

Rewrote add-cross and build-cross-gcc to work for multiple targets.

Moved handling of gcc flags from shell script into the compiler
proper.  This was done because we need different sets of flags for
different platforms.  It shouldn't cause any problems, since you can
always use -cc-opt to override them.

Eliminated Machine.{SetExnStackLocal,SetExnStackSlot,SetSlotExnStack},
which were vestigal.  They had been replaced by moves in the backend
a while ago.

Added -align-doubles {no|pad|skip}.  For now, only -align-doubles no
is implemented.  On SPARCs, which give a bus error for misaligned
doubles, this requires the C codegen to treat memory accesses to
doubles as two word accesses, which of course slows stuff down.  In
the near future, I plan to implement pad and skip.

Rewrote many of the switch usage messages so that <> is used to
indicate a variable that should be substituted for.

Eliminated -D.  You can use instead: -cc-opt '-D<SYM>'

Changed -build-constants to -build-constants {false|true}, for
uniformity's sake.