MAIL double alignment
Added mlton flag -align {4|8}.  With -align n, the following are
true:

	The first data word of objects, stacks, and stack frames is
	aligned mod n.

	The size of stack frames and object is divisible by n.
	object width

So, -align 4 is the same as we always used to have, and -align 8 is
new.  In addition, with -align 8, doubles stored in memory are
guaranteed to be aligned mod 8.  This follows from the fact the
offsets used for doubles in objects and stack slots are aligned mod 8,
plus the fact that the start of the object or stack frame is aligned
mod 8.

The compiler changes were to the representation pass, which now
ensures that object sizes are properly aligned and to register
allocation, which ensures that stack frame sizes are properly
aligned.

Fixed one hole in the MACHINE type checker.  It now checks that the
stack offset at which a label (for a Cont, Creturn, or Handler) is
stored corresponds to the frame size associated with the label.  Of
course, I found this by introducing a bug where this wasn't the case
and watching the stack pop of the wrong frame size put the stackTop at
the wrong place.

Most of the changes were in the runtime, to make sure that it properly
aligns the frontier.  Nothing very deep, just lots of details to get
the invariant right.  The biggest change had to do with crossMaps.
Since with -align 8, the header word is always guaranteed to be at 4
mod 8, the crossMap used by generational GC becomes useless.  That is,
it will never report that an object start coincides with a card start.
So, I changed the meaning of crossmap and improved its usage.  Instead
of having a single bit stored in a byte, I now use the entire byte to
record the start of the last object in card (or 255 if there is no
object start in the card).  This last object can be used as the start
point to scan the next marked card.

With -debug true, mlton now passes -DASSERT=1 to gcc when compiling C
files.  This works nicely with asserts in ccodegen.h.

I worked for a bit on doing selective alignment by frontier skipping,
i.e., only aligning objects with doubles in them by aligning the
frontier immediately before allocation of an object with a double.
Unfortunately, this doesn't work well with copying GC, since depending
on the vagaries of object order, more or less extra space might be
needed for frontier alignment.  So a copying GC might need more space
in to space than in from space.  I decided I didn't want to deal with
that, so I stuck with the alignment via object padding above.